+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/queue/ProcessQueueWatchdog.java
@@ -162,25 +162,35 @@ public class ProcessQueueWatchdog implements ScheduledTask {
         @Override
         public void run() {
             Field<OffsetDateTime> maxAge = PgUtils.nowMinus(cfg.getMaxFailureHandlingAge());
+            boolean doWork = true;
 
             for (PollEntry e : POLL_ENTRIES) {
                 List<ProcessEntry> parents = watchdogDao.poll(e, maxAge, 1);
 
                 for (ProcessEntry parent : parents) {
-                    process(e, parent);
+                    doWork = process(e, parent);
+                    if (!doWork) {
+                        break;
+                    }
+                }
+
+                if (!doWork) {
+                    break;
                 }
             }
         }
 
-        private void process(PollEntry entry, ProcessEntry parent) {
-            String username = userDao.getUsername(parent.initiatorId);
-
+        /**
+         * @return {@code true} on success, {@code false} on error
+         */
+        private boolean process(PollEntry entry, ProcessEntry parent) {
             Map<String, Object> req = new HashMap<>();
             req.put(Constants.Request.ENTRY_POINT_KEY, entry.flow);
             req.put(Constants.Request.TAGS_KEY, null); // clear tags
 
             PartialProcessKey childKey = PartialProcessKey.create();
             try {
+                String username = assertUserEnabled(parent.initiatorId);
                 Payload payload = payloadManager.createFork(childKey, parent.processKey, entry.handlerKind,
                         parent.initiatorId, username, parent.projectId, req, null,
                         null, parent.imports);
@@ -191,12 +201,31 @@ public class ProcessQueueWatchdog implements ScheduledTask {
 
                 log.info("process -> created a new child process '{}' (parent '{}', entryPoint: '{}')",
                         childKey, parent.processKey, entry.flow);
+                return true;
             } catch (Exception e) {
                 // remove the handler from the parent process to avoid infinite retries
                 queueDao.removeHandler(parent.processKey, entry.flow);
                 logManager.warn(parent.processKey, "Error while starting {} handler: {}", entry.flow, e.getMessage());
-                throw new RuntimeException(e);
+                log.warn("Error while starting {}/{} handler: {}", parent.processKey, entry.flow, e.getMessage());
+                return false;
             }
+        }
+
+        /**
+         * @return username
+         */
+        private String assertUserEnabled(UUID initiatorId) {
+            var user = userDao.get(initiatorId);
+
+            if (user == null) {
+                throw new IllegalStateException("initiator not found");
+            }
+
+            if (user.isDisabled()) {
+                throw new IllegalArgumentException("initiator is disabled");
+            }
+
+            return user.getName();
         }
 
         private String toString(ProcessKind kind) {