+++ server/impl/src/main/java/com/walmartlabs/concord/server/repository/RepositoryManager.java
@@ -73,6 +73,7 @@ public class RepositoryManager {
                 .httpLowSpeedTime(gitCfg.getHttpLowSpeedTime())
                 .sshTimeout(gitCfg.getSshTimeout())
                 .sshTimeoutRetryCount(gitCfg.getSshTimeoutRetryCount())
+                .maxGitCliOutputBytes(gitCfg.maxGitCliOutputBytes())
                 .build();
 
         this.gitCfg = gitCfg;